mcp/garmin/requirements.txt
mcp>=1.0.0
garminconnect>=0.2.0
python-dotenv>=1.0.0
